Change bungee message packet id to 249
