+++ src/main/java/net/minecraft/server/Packet.java
@@ -244,7 +244,7 @@ public abstract class Packet {
         a(255, true, true, Packet255KickDisconnect.class);
 
         // Poseidon start - custom packets
-        a(250, true, true, Packet250BungeePayload.class);
+        a(249, true, true, Packet249BungeePayload.class);
         packetClassToIdMap.put(ArtificialPacket53BlockChange.class, 53); //Poseidon - Artificial Block Change Packet
         // Poseidon end
